backend/pyproject.toml
[project]
name = "agent-platform-backend"
version = "0.0.1"
description = "Agent Platform MVP Backend"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "sqlalchemy>=2.0.0",
    "aiosqlite>=0.19.0",
    "pydantic>=2.5.0",
    "anthropic>=0.18.0",
    "httpx>=0.26.0",
    "python-dotenv>=1.0.0",
    "claude-agent-sdk>=0.1.18",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.26.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
